refactor: seperation of styled components from components for pages
