+++ src/pages/EventsPage/EventsPage.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 function EventsPage() {
   return <div>EventsPage</div>;
 }
