+++ src/pages/FaqPage/FaqPage.styled.ts
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+export const Container = styled.div`
+  h2 {
+    margin-bottom: 20px;
+  }
+  max-width: 1200px;
+  margin: 50px auto;
+`;
+
+export const styles = {
+  bgColor: "white",
+  titleTextColor: "rgb(29, 96, 157)",
+  rowTitleColor: "#444",
+  rowContentColor: "grey",
+  rowContentPaddingBottom: "50px",
+  // arrowColor: 'red',
+};
