+++ src/pages/MainRouter.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect } from "react";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 
-import IndexPage from "./IndexPage";
-import EventsPage from "./EventsPage";
-import AboutPage from "./AboutPage";
+import IndexPage from "./IndexPage/IndexPage";
+import EventsPage from "./EventsPage/EventsPage";
+import AboutPage from "./AboutPage/AboutPage";
 import SponsorsPage from "./SponsorsPage";
-import FaqPage from "./FaqPage";
+import FaqPage from "./FaqPage/FaqPage";
 
 function Redirect({ redirectUrl }: { redirectUrl: string }) {
   useEffect(() => {
